+++ backend/src/main/java/finki/db/tasty_tabs/entity/exceptions/TableNotFoundException.java
@@ -2,7 +2,7 @@ package finki.db.tasty_tabs.entity.exceptions;
 
 public class TableNotFoundException extends DomainException {
 
-    public TableNotFoundException(Integer id) {
+    public TableNotFoundException(Long id) {
         super(String.format("Table with id %d doesnt exist",id));
     }
 }
