+++ backend/src/main/java/finki/db/tasty_tabs/entity/exceptions/TableNumberAlreadyExistsException.java
@@ -2,7 +2,7 @@ package finki.db.tasty_tabs.entity.exceptions;
 
 public class TableNumberAlreadyExistsException extends DomainException {
 
-    public TableNumberAlreadyExistsException(Integer number) {
+    public TableNumberAlreadyExistsException(Long number) {
         super(String.format("Table: %d already exists", number));
     }
 }
