+++ backend/src/main/java/finki/db/tasty_tabs/service/ReservationService.java
@@ -3,7 +3,6 @@ package finki.db.tasty_tabs.service;
 import finki.db.tasty_tabs.entity.Reservation;
 import finki.db.tasty_tabs.entity.ReservationManagedFrontStaff;
 import finki.db.tasty_tabs.web.dto.CreateReservationDto;
-import finki.db.tasty_tabs.web.dto.ReservationDto;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -15,7 +14,7 @@ public interface ReservationService {
     Reservation updateReservation(Long id, CreateReservationDto dto,String userEmail);
     void deleteReservation(Long id);
     List<Reservation> getAllReservationsByUser(String userEmail);
-    ReservationManagedFrontStaff acceptReservation(Long reservationId, String frontStaffEmail, Integer tableNumber);
+    ReservationManagedFrontStaff acceptReservation(Long reservationId, String frontStaffEmail, Long tableNumber);
 
     List<Reservation> getAllReservationsForToday();
     List<Reservation> getAllReservationsForDate(LocalDate date);
