+++ backend/src/main/java/finki/db/tasty_tabs/service/RestaurantTableService.java
@@ -3,14 +3,13 @@ package finki.db.tasty_tabs.service;
 import finki.db.tasty_tabs.entity.RestaurantTable;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface RestaurantTableService {
 
-    RestaurantTable findById(Integer id);
+    RestaurantTable findById(Long id);
     List<RestaurantTable> getAll();
-    RestaurantTable updateTable(Integer id,RestaurantTable restaurantTable);
-    void deleteTable(Integer id);
+    RestaurantTable updateTable(Long id, RestaurantTable restaurantTable);
+    void deleteTable(Long id);
     List<RestaurantTable> getAllBySeatCapacity(Integer seatCapacity);
     RestaurantTable createTable(RestaurantTable restaurantTable);
 }
