+++ backend/src/main/java/finki/db/tasty_tabs/web/dto/CreateRestaurantTableDto.java
@@ -3,7 +3,7 @@ package finki.db.tasty_tabs.web.dto;
 import finki.db.tasty_tabs.entity.RestaurantTable;
 
 public record CreateRestaurantTableDto(
-        Integer tableNumber,
+        Long tableNumber,
         Integer seatCapacity
 ) {
     public RestaurantTable toRestaurantTable(){
