+++ backend/src/main/java/finki/db/tasty_tabs/web/dto/RestaurantTableDto.java
@@ -3,7 +3,7 @@ package finki.db.tasty_tabs.web.dto;
 import finki.db.tasty_tabs.entity.RestaurantTable;
 
 public record RestaurantTableDto(
-        Integer tableNumber,
+        Long tableNumber,
         Integer seatCapacity
 ) {
     public static RestaurantTableDto from(RestaurantTable table){
